WIP. Start to most of the pipeline into mode-specific classes.
This works well, and the VisualizationModeCallback properly handles
a change between pipeline classes (or in this test case,
choosing to render or not render the output of the
BoneOnlyPipeline. This facilitates the selective rendering of
different branches of the pipeline by placing each branch in
a class of this type.

However, a lot of forward declaration was necessary to avoid
build errors, presumably to some sort of issue with the ordering
of included header files, or something similar. Must fix soon.